fix: move retry-on-conflict into LatencyHistory, remove StorableWriteWithRetry
StorableWriteWithRetry was a general-purpose API for a pattern that is
only correct for commutative mutations (list appends). Exposing it as
reusable invited misuse on non-commutative operations where last-retrier-
wins drops concurrent mutations.

Inline the retry loop as LatencyHistory.WriteWithRetry — a private method
on the one type where the pattern is correct. Remove the exported API,
its constants, and its helper from evr_storable.go.

Tests moved from evr_storable_retry_test.go into evr_latencyhistory_test.go
alongside the existing LatencyHistory tests.
