+++ server/evr_storable.go
@@ -4,40 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"math/rand"
 	"reflect"
-	"strings"
-	"time"
 
 	"github.com/heroiclabs/nakama-common/runtime"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
-
-const (
-	// storableMaxWriteAttempts bounds StorableWriteWithRetry so a persistently
-	// contended object cannot spin forever. Includes the initial attempt.
-	storableMaxWriteAttempts = 5
-	// storableRetryBaseBackoff is the base unit of jittered backoff between
-	// retry attempts. Kept small: contention here is brief (per-user object).
-	storableRetryBaseBackoff = 5 * time.Millisecond
-)
-
-// isStorableVersionConflict reports whether err is a storage optimistic-
-// concurrency (version-check) conflict, as opposed to any other failure.
-//
-// We match on the version-conflict signature specifically. Nakama's
-// StorageWrite returns runtime.ErrStorageRejectedVersion ("Storage write
-// rejected - version check failed.") for OCC conflicts; StorableWrite wraps
-// that with %v, so the sentinel identity is lost but its message survives.
-// Matching the substring keeps detection working through that wrapper and
-// mirrors the existing isVersionConflictError helper in this package.
-func isStorableVersionConflict(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(err.Error(), "version check failed")
-}
 
 // StorageObjectAdapter defines methods for converting to a storage object.
 type StorableAdapter interface {
@@ -163,58 +135,4 @@ func StorableWrite(ctx context.Context, nk runtime.NakamaModule, userID string, 
 		src.SetStorageMeta(meta)
 	}
 	return nil
-}
-
-// StorableWriteWithRetry writes src, retrying on optimistic-concurrency
-// (version-check) conflicts only. On such a conflict it re-reads the current
-// stored object into src (adopting the concurrent winner's version and
-// contents), invokes reapply to re-apply this caller's pending mutation onto
-// that fresh object, and writes again. This makes concurrent writers lossless:
-// neither the winner's nor this caller's new entries are clobbered.
-//
-// reapply MUST be idempotent across attempts and MUST re-apply the same pending
-// mutation onto the (now-refreshed) src each time it is called. It is NOT
-// invoked before the first attempt — the caller is responsible for having
-// applied its mutation to src once before calling. Any pruning/limit/expiry
-// logic that must stay bounded after a merge should live inside reapply.
-//
-// Only version-conflict errors are retried; any other error (including
-// non-conflict storage failures) is returned immediately without re-reading.
-// Attempts are bounded by storableMaxWriteAttempts with small jittered backoff.
-func StorableWriteWithRetry(ctx context.Context, nk runtime.NakamaModule, userID string, src StorableAdapter, reapply func() error) error {
-	var lastErr error
-	for attempt := 0; attempt < storableMaxWriteAttempts; attempt++ {
-		err := StorableWrite(ctx, nk, userID, src)
-		if err == nil {
-			return nil
-		}
-		if !isStorableVersionConflict(err) {
-			// Not an OCC conflict: do not re-read or retry.
-			return err
-		}
-		lastErr = err
-
-		// Conflict: a concurrent writer advanced the version. Re-read the fresh
-		// object (picking up the winner's version + entries) then re-apply this
-		// caller's pending mutation onto it before writing again.
-		if rerr := StorableRead(ctx, nk, userID, src, false); rerr != nil {
-			return fmt.Errorf("StorableWriteWithRetry: re-read after conflict: %w", rerr)
-		}
-		if rerr := reapply(); rerr != nil {
-			return fmt.Errorf("StorableWriteWithRetry: re-apply after conflict: %w", rerr)
-		}
-
-		// Jittered backoff before the next attempt (skip after the final loop).
-		if attempt < storableMaxWriteAttempts-1 {
-			backoff := time.Duration(attempt+1) * storableRetryBaseBackoff
-			// rand is fine here: this is contention jitter, not security.
-			jitter := time.Duration(rand.Int63n(int64(storableRetryBaseBackoff)))
-			select {
-			case <-ctx.Done():
-				return fmt.Errorf("StorableWriteWithRetry: %w", ctx.Err())
-			case <-time.After(backoff + jitter):
-			}
-		}
-	}
-	return fmt.Errorf("StorableWriteWithRetry: exhausted %d attempts: %w", storableMaxWriteAttempts, lastErr)
 }
